% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseSet-class}
\alias{BaseSet-class}
\title{BaseSet Class}
\description{
The \code{BaseSet} class implements a minimal container to describe distinct objects that make up sets.
}
\section{Slots}{

\describe{
\item{\code{map}}{DataFrame. Two columns provide mapping relationships between \code{"element"} and \code{"set"}.}

\item{\code{elementData}}{DataFrame. Provide metadata for each unique element in \code{map$element}.}

\item{\code{setData}}{DataFrame. Provide metadata for each unique element in \code{map$set}.}
}}

