% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{BaseSet}
\alias{BaseSet}
\title{Constructor for the \code{BaseSet} class}
\usage{
BaseSet(map, elementData, setData)
}
\arguments{
\item{map}{DataFrame. Two columns provide mapping relationships between \code{"element"} and \code{"set"}.}

\item{elementData}{DataFrame. Provide metadata for each unique element in \code{map$element}.}

\item{setData}{DataFrame. Provide metadata for each unique element in \code{map$set}.}
}
\value{
A \code{BaseSet} object.
}
\description{
Constructor for the \code{BaseSet} class
}
\examples{
# Visually intuitive definition of sets
sets <- list(
  set1=c("A", "B"),
  set2=c("C", "D"))

# Reformat as a table
map <- DataFrame(
  element=unlist(sets),
  set=rep(names(sets), lengths(sets))
)

bs <- BaseSet(map)
}
