% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R,
%   R/FuzzyHits-class.R
\docType{methods}
\name{FuzzyHits-methods}
\alias{FuzzyHits-methods}
\alias{membership}
\alias{membership<-}
\alias{membership,FuzzyHits-method}
\alias{membership<-,FuzzyHits-method}
\title{Methods for \code{FuzzyHits} Objects}
\usage{
membership(object)

membership(object) <- value

\S4method{membership}{FuzzyHits}(object)

\S4method{membership}{FuzzyHits}(object) <- value
}
\arguments{
\item{object}{An object of class inheriting from \link[=FuzzyHits-class]{FuzzyHits}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Accessors'.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\link[=FuzzyHits-class]{FuzzyHits}} class.
The \code{FuzzyHits} class directly extends \code{\link[=Hits-class]{Hits}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{x} represents an object of class inheriting from \code{\link[=FuzzyHits-class]{FuzzyHits}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{membership(object)} returns a \code{numeric} vector of membership function for each relation.
}

\examples{
# Setup ----

example("FuzzyHits-class", echo=FALSE)

# Accessors ----

membership(fh)
fh1 <- fh
membership(fh1)[1] <- 0
}
\seealso{
\code{\linkS4class{FuzzyHits}}, \linkS4class{FuzzySets}.
}
