% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/FuzzySet-class.R
\docType{class}
\name{FuzzySets-class}
\alias{FuzzySets-class}
\alias{FuzzySets}
\alias{membership}
\alias{membership<-}
\alias{membership,FuzzySets-method}
\alias{membership<-,FuzzySets-method}
\alias{as.list.FuzzySets}
\alias{as.matrix.FuzzySets}
\title{FuzzySets Class}
\usage{
FuzzySets(relations = DataFrame(element = character(0), set =
  character(0), membership = numeric(0)), ...)

membership(x)

membership(x) <- value

\S4method{membership}{FuzzySets}(x)

\S4method{membership}{FuzzySets}(x) <- value

\method{as.list}{FuzzySets}(x, ...)

\method{as.matrix}{FuzzySets}(x, ..., fill = NA_real_)
}
\arguments{
\item{relations}{DataFrame. At least 3 columns provide mapping relationships between \code{"element"} and \code{"set"} with \code{"membership"} function in the range \code{[0,1]}.}

\item{...}{Arguments passed to the \code{\link[=BaseSets]{BaseSets()}} constructor and other functions.}

\item{x}{An object that inherits from \code{FuzzySets}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Slots'.}

\item{fill}{Value with which to fill in structural missings, defaults to 0 (element is not a member of the set).}
}
\value{
A \code{FuzzySets} object.
}
\description{
The \code{FuzzySets} class extends the \code{\link{BaseSets}} class to implement a container that also describe different grades of membershipin the interval \code{[0,1]}.
}
\section{Conversion to matrix}{

As it is possible to store multiple relations between the same gene and gene set, it may be necessary to collapse multiple observations of the membership function into a single value.
This can be controlled using the \code{fun.aggregate} argument passed down to the \code{acast} function.
See examples below.
}

\examples{
# Constructor ----

# Visually intuitive definition of sets, elements, and membership
sets <- list(
  set1=c("A"=0.1, "B"=0.2),
  set2=c("B"=0.3, "C"=0.4, "D"=0.5),
  set3=c("E"=0.8))

# unlist the set names
unlistSets <- rep(names(sets), lengths(sets))
# unlist the element names
unlistElements <- unlist(sapply(sets, names))
# unlist the membership values
unlistMembership <- unlist(sets)

# Reformat as a table
relations <- DataFrame(
  element=unlistElements,
  set=unlistSets,
  membership=unlistMembership
)

fs <- FuzzySets(relations=relations)

# Subsetting ----

fs1 <- subset(fs, set == "set1" | membership > 0.5)

# Coercing ----

# to list (gene sets)
ls1 <- as(fs, "list")
# to matrix (continuous membership)
ls1 <- as(fs, "matrix")

# Getters/Setters ----

membership(fs)

fs1 <- fs
membership(fs1) <- runif(length(fs1))

# Converting to matrix: multiple observations ----

as.matrix(fs, fun.aggregate=min)

}
\seealso{
\code{\link{BaseSets}}

\code{\link{BaseSets}}
}
