% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/FuzzySets-class.R
\docType{methods}
\name{FuzzySets-methods}
\alias{FuzzySets-methods}
\alias{membership,FuzzySets-method}
\alias{membership<-,FuzzySets-method}
\alias{subset.FuzzySets}
\alias{subset,FuzzySets-method}
\alias{as.matrix.FuzzySets}
\alias{as.FuzzySets.matrix}
\alias{as.FuzzySets}
\title{Methods for \code{FuzzySets} Objects}
\usage{
\S4method{membership}{FuzzySets}(x)

\S4method{membership}{FuzzySets}(x) <- value

\method{subset}{FuzzySets}(x, ...)

\method{as.matrix}{FuzzySets}(x, fill = NA_real_, ...)

as.FuzzySets.matrix(matrix, ...)
}
\arguments{
\item{x}{An object of class inheriting from \code{\link[=FuzzySets-class]{FuzzySets}}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Accessors'.}

\item{...}{Additional arguments passed to and from other methods.}

\item{fill}{Value with which to fill in structural missings, passed to \code{\link[=acast]{acast()}}.
Defaults to \code{NA_real_}, to contrast with relations explictly associated with a membership function of 0.}

\item{matrix}{A \code{matrix}.
The matrix will be coerced to \code{double} type and the value will be taken to indicate the membership function.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\link[=FuzzySets-class]{FuzzySets}} class.
The \code{FuzzySets} class directly extends \code{\link[=BaseSets-class]{BaseSets}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{x} represents an object of class inheriting from \code{\link[=FuzzySets-class]{FuzzySets}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{membership(x)} returns a \code{numeric} vector of membership function for each relation.
}

\section{Subsetting}{

\code{subset(x, subset, ...)} returns subsets of relations which meet conditions.
For \code{FuzzySets} objects, the \code{subset} argument should be a logical expression referring to any of \code{"element"}, \code{"set"}, \code{"membership"}, and any available relation metadata indicating elements or rows to keep: missing values are taken as false.
In addition, metadata for elements and sets that are not represented in the remaining relations are also dropped.
}

\section{Coercion}{

\code{as(x, "matrix")} and \code{as.matrix(x)} return a \code{matrix} with elements as rows, sets as columns, and a numeric value indicating the membership function.
}

\section{Coercion to matrix}{

As it is possible to store multiple relations between the same gene and gene set, it may be necessary to collapse multiple observations of the membership function into a single value.
To this end, the form \code{as.matrix(x, fun.aggregate)} can be used to provide an aggregation function.
See examples.
}

\examples{
# Setup ----

example("FuzzySets-class", echo=FALSE)

# Accessors ----

membership(fs)
fs1 <- fs
membership(fs1)[1] <- 0

# Subsetting ----

fs1 <- subset(fs, set == "set1" | membership > 0.5)

# Coercion from/to FuzzySets ----

matrix1 <- as(fs, "matrix")

# Coercion to FuzzySets ----

fs1 <- as(matrix1, "FuzzySets")
}
\seealso{
\code{\linkS4class{FuzzySets}}, \code{\link{BaseSets-methods}}.
}
