% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{data}
\name{GOHits-class}
\alias{GOHits-class}
\alias{GOEvidenceCodes}
\alias{GOOntologyCodes}
\alias{GOHits}
\title{GOHits Class}
\format{A named vector of length 26.
\describe{
\item{names}{Code.}
\item{carat}{Description.}
...
}}
\source{
\url{http://geneontology.org/docs/guide-go-evidence-codes/}

\url{http://geneontology.org/docs/guide-go-evidence-codes/}
}
\usage{
GOEvidenceCodes

GOOntologyCodes

GOHits(from = integer(0), to = integer(0),
  evidence = factor(character(0), names(GOEvidenceCodes)),
  ontology = factor(character(0), names(GOOntologyCodes)), nLnode = 0L,
  nRnode = 0L, ...)
}
\arguments{
\item{from, to}{Two integer vectors of the same length.
The values in \code{from} must be >= 1 and <= \code{nLnode}.
The values in \code{to} must be >= 1 and <= \code{nRnode}.}

\item{evidence}{factor. Levels must be values in \code{names(GOEvidenceCodes)}.}

\item{ontology}{factor. Levels must be values in \code{names(GOOntologyCodes)}.}

\item{nLnode, nRnode}{Number of left and right nodes.}

\item{...}{Arguments metadata columns to set on the \code{GOHits} object.
All the metadata columns must be vector-like objects of the same length as \code{from}, \code{to}, and \code{membership}.}
}
\value{
A \code{GOHits} object.
}
\description{
The \code{GOHits} class extends the \code{\linkS4class{Hits}} class to represent hits that also describe relations between genes and sets using the Gene Ontology controlled vocabulary.
}
\details{
This class does not define any additional slot to the \code{Hits} class.
However, this class defines additional validity checks to ensure that every relation stored in a \code{GOHits} are respect the Gene Ontology evidence and ontology codes.
Refer to \code{\link{GOOntologyCodes}} and \code{\link{GOEvidenceCodes}} for valid code and vocabulary.
}
\section{Controlled vocabulary}{

Gene Ontology evidence codes were obtained from <>


Gene Ontology namespaces were obtained from \url{http://geneontology.org/docs/ontology-documentation/}
}

\examples{
# Controlled vocabulary ----

GOEvidenceCodes
GOOntologyCodes

# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
ontology <- factor(c("BP", "BP", "BP", "MF", "MF", "CC"))
evidence <- factor(c("IEA", "IDA", "IEA", "IDA", "IEA", "IDA"))

gh <- GOHits(from, to, evidence, ontology, 7, 15)
gh
}
\seealso{
\code{\linkS4class{Hits}}, \code{\linkS4class{FuzzySets}}
}
\author{
Kevin Rue-Albrecht
}
\keyword{datasets}
