% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R, R/GOHits-class.R
\docType{methods}
\name{GOHits-methods}
\alias{GOHits-methods}
\alias{ontology}
\alias{ontology<-}
\alias{evidence}
\alias{evidence<-}
\alias{evidence,GOHits-method}
\alias{evidence<-,GOHits-method}
\alias{ontology,GOHits-method}
\alias{ontology<-,GOHits-method}
\title{Methods for \code{GOHits} Objects}
\usage{
ontology(object)

ontology(object) <- value

evidence(object)

evidence(object) <- value

\S4method{evidence}{GOHits}(object)

\S4method{evidence}{GOHits}(object) <- value

\S4method{ontology}{GOHits}(object)

\S4method{ontology}{GOHits}(object) <- value
}
\arguments{
\item{object}{An object of class inheriting from \link[=GOHits-class]{GOHits}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Accessors'.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\link[=GOHits-class]{GOHits}} class.
The \code{GOHits} class directly extends \code{\link[=Hits-class]{Hits}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{object} represents an object of class inheriting from \code{\link[=GOHits-class]{GOHits}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{evidence(object)} returns a \code{factor} indiating the evidence code for each relation.


\code{ontology(object)} returns a \code{numeric} vector of ontology function for each relation.
}

\examples{
# Controlled vocabulary ----

GOEvidenceCodes
GOOntologyCodes
# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
ontology <- factor(c("BP", "BP", "BP", "MF", "MF", "CC"))
evidence <- factor(c("IEA", "IDA", "IEA", "IDA", "IEA", "IDA"))

gh <- GOHits(from, to, ontology, evidence, 7, 15)
gh

# Accessors ----

evidence(gh)
gh1 <- gh
evidence(gh1)[1] <- "HTP"
ontology(gh)
gh1 <- gh
ontology(gh1)[1] <- "MF"
}
\seealso{
\code{\linkS4class{GOHits}}, \linkS4class{GOHits}.
}
\author{
Kevin Rue-Albrecht
}
