% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GOSets-class}
\alias{GOSets-class}
\alias{GOSets}
\title{GOSets Class}
\usage{
GOSets(relations = DataFrame(element = character(0), set = character(0),
  ontology = factor(character(0), names(GOOntologyCodes)), evidence =
  factor(character(0), names(GOEvidenceCodes))), ...)
}
\arguments{
\item{relations}{\code{\linkS4class{DataFrame}}.
At least 3 columns that provide mapping relationships between \code{"element"} and \code{"set"}, with \code{"membership"} function in the range \code{[0,1]}.
Additional columns are taken as relation metadata.}

\item{...}{Arguments passed to the \code{\link[=Sets]{Sets()}} constructor and other functions.}
}
\value{
A \code{GOSets} object.
}
\description{
The \code{GOSets} class extends the \code{\linkS4class{Sets}} class to implement a container that also describes relations between genes and sets using the Gene Ontology controlled vocabulary.
Refer to \code{\link{GOOntologyCodes}} and \code{\link{GOEvidenceCodes}} for valid vocabulary.
}
\examples{
# Constructor ----

# Fetch a sample of GO annotations
library(org.Hs.eg.db)
go_sets <- import(org.Hs.egGO)
relations <- as.data.frame(head(go_sets))

gs <- GOSets(relations)

# Subsetting ----

gs1 <- subset(gs, element == "1" & ontology == "BP" & evidence == "TAS")
relations(gs1)

# Getters/Setters ----

evidence(gs)
ontology(gs)

gs1 <- gs
evidence(gs1)[1] <- "EXP"

gs1 <- gs
ontology(gs1)[1] <- "CC"
}
\seealso{
\code{\linkS4class{Sets}}, \code{\linkS4class{GOHits}}, \code{\link{GOSets-methods}}.
}
\author{
Kevin Rue-Albrecht
}
