% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/GOSets-class.R
\docType{methods}
\name{GOSets-methods}
\alias{GOSets-methods}
\alias{evidence,GOSets-method}
\alias{evidence<-,GOSets-method}
\alias{ontology,GOSets-method}
\alias{ontology<-,GOSets-method}
\alias{subset.GOSets}
\alias{subset,GOSets-method}
\title{Methods for \code{GOSets} Objects}
\usage{
\S4method{evidence}{GOSets}(object)

\S4method{evidence}{GOSets}(object) <- value

\S4method{ontology}{GOSets}(object)

\S4method{ontology}{GOSets}(object) <- value

\method{subset}{GOSets}(x, ...)
}
\arguments{
\item{object, x}{An object of class inheriting from \code{\linkS4class{GOSets}}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Accessors'.}

\item{...}{Additional arguments passed to and from other methods.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\linkS4class{GOSets}} class.
The \code{GOSets} class directly extends \code{\linkS4class{BaseSets}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{object} and \code{x} represent an object of class inheriting from \code{\linkS4class{GOSets}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{evidence(object)} returns a \code{numeric} vector of evidence function for each relation.


\code{ontology(object)} returns a \code{numeric} vector of ontology function for each relation.
}

\section{Subsetting}{

\code{subset(x, subset, ...)} returns subsets of relations which meet conditions.
For \code{GOSets} objects, the \code{subset} argument should be a logical expression referring to any of \code{"element"}, \code{"set"}, \code{"evidence"}, \code{"ontology"} and other available relation metadata indicating elements or rows to keep: missing values are taken as false.
In addition, metadata for elements and sets that are not represented in the remaining relations are also dropped.
}

\examples{
# Constructor ----

# Fetch a sample of GO annotations
library(org.Hs.eg.db)
base_sets <- import(org.Hs.egGO)
relations <- as.data.frame(head(base_sets))

gs <- GOSets(relations)

# Accessors ----

evidence(gs)
gs1 <- gs
evidence(gs1)[1] <- "EXP"
ontology(gs)
gs1 <- gs
ontology(gs1)[1] <- "CC"

# Subsetting ----

gs1 <- subset(gs, ontology == "BP" & evidence == "TAS")
gs1
}
\seealso{
\code{\linkS4class{GOSets}}, \code{\link{BaseSets-methods}}.
}
\author{
Kevin Rue-Albrecht
}
