% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R,
%   R/IdVector-class.R
\docType{methods}
\name{IdVector-methods}
\alias{IdVector-methods}
\alias{ids}
\alias{ids<-}
\alias{ids,IdVector-method}
\alias{ids<-,IdVector-method}
\alias{names,IdVector-method}
\alias{names<-,IdVector-method}
\alias{length,IdVector-method}
\alias{[,IdVector,ANY,ANY,ANY-method}
\alias{[,IdVector-method}
\alias{as.vector.IdVector}
\alias{as.vector}
\alias{as.character.IdVector}
\alias{as.character}
\alias{as.IdVector.default}
\alias{as.IdVector}
\title{Methods for \code{IdVector} objects}
\usage{
ids(x)

ids(x) <- value

\S4method{ids}{IdVector}(x)

\S4method{ids}{IdVector}(x) <- value

\S4method{names}{IdVector}(x)

\S4method{names}{IdVector}(x) <- value

\S4method{length}{IdVector}(x)

\S4method{[}{IdVector,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)

\method{as.vector}{IdVector}(x, mode = "character")

\method{as.character}{IdVector}(x, ...)

as.IdVector.default(ids, ...)
}
\arguments{
\item{x}{An object of class inheriting from \link[=IdVector-class]{IdVector}.}

\item{value}{An object of a class specified in the S4 method signature.}

\item{i}{index specifying elements to extract or replace.}

\item{j, ..., drop}{Ignored.}

\item{mode}{Ignored. The vector will be coerced to \code{character} mode, unless requested otherwise.}

\item{ids}{An atomic vector of identifiers.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\link[=IdVector-class]{IdVector}} class.
The \code{IdVector} class directly extends \code{\link[=Vector-class]{Vector}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{x} represents an object of class inheriting from \code{\link[=IdVector-class]{IdVector}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{ids(x)} returns a \code{character} vector of element identifiers.
\code{names(x)} is a synonym for compatibility with S4 methods such as \code{mcols(x, use.names = TRUE, ...) }.
}

\section{Dimensions}{

\code{length(x)} returns the number of elements in \code{x}.
}

\section{Subsetting}{

\code{x[i]} returns new \code{\link[=IdVector-class]{IdVector}} object of the same class as \code{x} made of the elements selected by \code{i}. \code{i} can be missing; an NA-free logical, numeric, or character vector or factor (as ordinary vector or \code{\link{Rle}} object); or an \code{\link[=IntegerRanges-class]{IntegerRanges}} object.
}

\section{Coercion}{

\code{as(x, "vector")} and \code{as.vector(x)} return an atomic vector of identifiers contained in \code{x}.


\code{as(x, "vector")} and \code{as.vector(x)} return a \code{character} vector of identifiers contained in \code{x}.


\code{as(x, "IdVector")} and \code{as.IdVector(x)} return an \code{IdVector} from the given atomic vector of identifiers.
}

\examples{
# Setup ----

example("IdVector-class", echo=FALSE)

# Accessors ----

ids(iv)
iv1 <- iv
ids(iv1)[1] <- "gene1"

names(iv)
iv1 <- iv
names(iv1)[1] <- "GENE001"

# Dimensions ----

length(iv)

# Subsetting ----

iv1 <- iv[1:5]

# Coercion from IdVector ----

v1 <- as(iv, "vector")
c1 <- as(iv, "character")
}
\seealso{
\code{\linkS4class{IdVector}}, \code{\link[=Vector-class]{Vector}}, \code{\link{Vector-setops}}.
}
