% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\name{importGMT}
\alias{importGMT}
\title{Import GMT File}
\usage{
importGMT(path)
}
\arguments{
\item{path}{a file name or URL containing gene sets.}
}
\value{
Returns a \code{BaseSets} object with slot \code{setData} containing set
source information.
}
\description{
Imports a \code{.gmt} format file and produces a \code{BaseSets} class object.
A \code{.gmt} file is a non-rectangular tab-separated format, where each row
delimits a unique geneset. The first column is the geneset name, the second
column is the geneset source (such as a URL), and the third columns onwards
contain the names of the geneset constituents, such that each geneset can
have a variable length.
}
\examples{
gmt_file <- system.file(package = "unisets", "extdata",
    "example.gmt")
gmt_BaseSet <- importGMT(gmtFile)
}
\author{
Robert Amezquita, adapted from Kayla Morrell GeneSe pkg
}
