% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\docType{class}
\name{io}
\alias{io}
\alias{GMTFile-class}
\alias{GMTFile}
\alias{import.gmt}
\alias{import}
\alias{import,GMTFile,ANY,ANY-method}
\alias{export.gmt}
\alias{export}
\alias{export,Sets,GMTFile,ANY-method}
\alias{import.Go3AnnDbBimap}
\alias{import,Go3AnnDbBimap,ANY,ANY-method}
\title{Import and export}
\usage{
GMTFile(resource)

import.gmt(con, ...)

\S4method{import}{GMTFile,ANY,ANY}(con, format, text, ...)

export.gmt(object, con, ...)

\S4method{export}{Sets,GMTFile,ANY}(object, con, format, ...)

import.Go3AnnDbBimap(con, format, text, ...)

\S4method{import}{Go3AnnDbBimap,ANY,ANY}(con, format, text, ...)
}
\arguments{
\item{resource}{For \code{GMTFile()}, the .gmt file that will be imported in.}

\item{con}{The connection from which data is loaded or to which data is saved.
If this is a character vector, it is assumed to be a filename
and a corresponding file connection is created and then closed after exporting the object.
If a \code{\linkS4class{RTLFile}} derivative, the data is loaded from or saved to the underlying resource.
Certain subclasses of \code{\link[=AnnDbBimap]{BiMap}} are supported: \code{\link{Go3AnnDbBimap}}.}

\item{...}{Parameters to pass to the format-specific method.}

\item{format, text}{Arguments defined in the \code{\link[rtracklayer:import]{rtracklayer::import()}} generic. Currently ignored.}

\item{object}{An object of class inheriting from \code{\link{GMTFile}}.}

\item{Go3AnnDbBimap}{A \code{\link{Go3AnnDbBimap}}.}
}
\value{
For \code{GMTFile()}, an object representing the path to a .gmt file on disk
}
\description{
The functions \code{import} and \code{export} load and save objects from and to particular file formats.
The \code{unisets} package aims to implement support for a number of annotation and sequence formats.
The following file formats are currently supported: \href{https://software.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#GMT:_Gene_Matrix_Transposed_file_format_.28.2A.gmt.29}{GMT}.
}
\section{Coercion to Sets}{

\code{as(Go3AnnDbBimap, "Sets")} and \code{as.Sets(Go3AnnDbBimap)} return a \code{Sets} from a Gene Ontology \code{Bimap} stored distributed in a Bioconductor annotation package.
}

\examples{
# Example file ----

gmt <- system.file(package="unisets", "extdata", "example.gmt")

# Import ----

bs <- import(gmt)
bs

# Export ----

tf <- tempfile(fileext = ".gmt")
export(bs, tf)

# Import (Go3AnnDbBimap) ----

library(org.Hs.eg.db)
bs1 <- import(org.Hs.egGO)
bs1
}
\author{
Robert A. Amezquita
}
