% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{Sets-class}
\alias{Sets-class}
\alias{Sets}
\title{Sets Class}
\usage{
Sets(relations = DataFrame(element = character(0), set = character(0)),
  elementInfo, setInfo)
}
\arguments{
\item{relations}{\code{\linkS4class{DataFrame}}.
At least two columns that provide mapping relationships between \code{"element"} and \code{"set"} identifiers.
Additional columns are taken as relation metadata.}

\item{elementInfo}{\code{\link{IdVector}}.
Metadata for each unique identifier in \code{relations$element} is provided as \code{mcols(elementInfo)}.}

\item{setInfo}{\code{\link{IdVector}}.
Metadata for each unique identifier in \code{relations$set} is provided as \code{mcols(setInfo)}.}
}
\value{
A \code{Sets} object.
}
\description{
The \code{Sets} class implements a container to describe distinct objects that make up sets, along with element metadata and set metadata.
}
\section{Slots}{

\describe{
\item{\code{relations}}{\code{\linkS4class{Hits}}
The \emph{left node} and \emph{right node} of each hit stores the index of the \code{element} and \code{set} in \code{elementInfo} and \code{setInfo}, respectively.
Metadata for each relation is stored as \code{mcols(relations(object))}.}

\item{\code{elementInfo}}{\code{\linkS4class{IdVector}}.
Metadata for each unique element in \code{relations$element} is stored as \code{mcols(elementInfo)}.}

\item{\code{setInfo}}{\code{\linkS4class{IdVector}}.
Metadata for each unique set in \code{relations$set} is stored as \code{mcols(setInfo)}.}
}}

\examples{
# Constructor ----

# Visually intuitive definition of sets
sets <- list(
  set1=c("A", "B"),
  set2=c("B", "C", "D"),
  set3=c("E"))

bs <- as(sets, "Sets")
bs

# Coercing ----

# to list (gene sets)
ls1 <- as(bs, "list")
ls1
# to matrix (logical membership)
m1 <- as(bs, "matrix")
m1

# Accessors ----

relations(bs)
elementInfo(bs)
setInfo(bs)

# Dimensions ----

length(bs)
nElements(bs)
nSets(bs)

setLengths(bs)
elementLengths(bs)
}
\seealso{
\code{\link{Sets-methods}}.
}
\author{
Kevin Rue-Albrecht
}
