% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset-package.R
\docType{package}
\name{uniset-package}
\alias{uniset}
\alias{uniset-package}
\title{uniset: What the Package Does (one line, title case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
