% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset-package.R
\docType{package}
\name{unisets-package}
\alias{unisets}
\alias{unisets-package}
\title{unisets: Collection of Classes to Store Gene Sets}
\description{
Classes to describe relationships between elements and sets,
    with an emphasis on gene sets.
    Slots are available to store element and set metadata.
    Fuzzy sets (including membership functions) are supported.
}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk}

Authors:
\itemize{
  \item Robert Amezquita
}

}
\keyword{internal}
