% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{BaseSet-class}
\alias{BaseSet-class}
\alias{BaseSet}
\title{BaseSet objects}
\value{
A `BaseSet` object.
}
\description{
The `BaseSet` class implements a minimal container to describe distinct objects that make up sets.
}
\section{Slots}{

\describe{
\item{\code{map}}{DataFrame.}
}}

\examples{
bs <- BaseSet()
}
