% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/BaseSet-class.R
\docType{class}
\name{BaseSet-class}
\alias{BaseSet-class}
\alias{BaseSet}
\alias{setLengths}
\alias{elementLengths}
\alias{as.list.BaseSet}
\alias{as.list}
\alias{setLengths,BaseSet-method}
\alias{elementLengths,BaseSet-method}
\title{BaseSet Class}
\usage{
BaseSet(map, elementData, setData)

setLengths(x)

elementLengths(x)

\method{as.list}{BaseSet}(x, ...)

\S4method{setLengths}{BaseSet}(x)

\S4method{elementLengths}{BaseSet}(x)
}
\arguments{
\item{map}{DataFrame. Two columns provide mapping relationships between \code{"element"} and \code{"set"}.}

\item{elementData}{DataFrame. Provide metadata for each unique element in \code{map$element}.}

\item{setData}{DataFrame. Provide metadata for each unique element in \code{map$set}.}

\item{x}{An object that inherits from \code{BaseSet}.}

\item{...}{Additional arguments passed to and from methods.}
}
\value{
A \code{BaseSet} object.
}
\description{
The \code{BaseSet} class implements a container to describe distinct objects that make up sets, along with element metadata and set metadata.
}
\section{Slots}{

\describe{
\item{\code{map}}{DataFrame. Two columns provide mapping relationships between \code{"element"} and \code{"set"}.}

\item{\code{elementData}}{DataFrame. Provide metadata for each unique element in \code{map$element}.}

\item{\code{setData}}{DataFrame. Provide metadata for each unique element in \code{map$set}.}
}}

\examples{
# Constructor ----

# Visually intuitive definition of sets
sets <- list(
  set1=c("A", "B"),
  set2=c("C", "D"),
  set3=c("E"))

# Reformat as a table
map <- DataFrame(
  element=unlist(sets),
  set=rep(names(sets), lengths(sets))
)

bs <- BaseSet(map)

# Subsetting ----

bs1 <- subset(bs, set == "set1" | element == "E")

# Coercing to list ----
ls1 <- as(bs, "list")
}
