% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/FuzzyHits-class.R
\docType{class}
\name{FuzzyHits-class}
\alias{FuzzyHits-class}
\alias{FuzzyHits}
\alias{membership,FuzzyHits-method}
\alias{membership<-,FuzzyHits-method}
\title{FuzzyHits Class}
\usage{
FuzzyHits(from = integer(0), to = integer(0), nLnode = 0L,
  nRnode = 0L, membership = numeric(0), ...)

\S4method{membership}{FuzzyHits}(x)

\S4method{membership}{FuzzyHits}(x) <- value
}
\arguments{
\item{from, to}{Two integer vectors of the same length.
The values in \code{from} must be >= 1 and <= \code{nLnode}.
The values in \code{to} must be >= 1 and <= \code{nRnode}.}

\item{nLnode, nRnode}{Number of left and right nodes.}

\item{membership}{Numeric. Vector of membership in the range \code{[0,1]}}

\item{...}{Arguments to pass to the \code{\link[=Hits]{Hits()}} constructor,
or to and from other functions.}

\item{x}{An object that inherits from \code{FuzzyHits}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Slots'.}
}
\value{
A \code{FuzzyHits} object.
}
\description{
The \code{FuzzyHits} class extends the \code{\link{Hits}} class to represent hits that are associated with different grades of membershipin the interval \code{[0,1]}.
}
\section{Slots}{

\describe{
\item{\code{membership}}{numeric. Membership function.}
}}

\examples{
# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
membership <- c(0, 0.1, 0.2, 0.3, 0.6, 0.8)

fh <- FuzzyHits(from, to, 7, 15, membership)

}
\seealso{
\code{\link{Hits}}, \code{\link{FuzzySets}}
}
