% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FuzzyHits-class}
\alias{FuzzyHits-class}
\alias{FuzzyHits}
\title{FuzzyHits Class}
\usage{
FuzzyHits(from = integer(0), to = integer(0),
  membership = numeric(0), nLnode = 0L, nRnode = 0L, ...)
}
\arguments{
\item{from, to}{Two integer vectors of the same length.
The values in \code{from} must be >= 1 and <= \code{nLnode}.
The values in \code{to} must be >= 1 and <= \code{nRnode}.}

\item{membership}{Numeric. Vector of numeric membership function in the range \code{[0,1]}}

\item{nLnode, nRnode}{Number of left and right nodes.}

\item{...}{Arguments metadata columns to set on the \code{FuzzyHits} object.
All the metadata columns must be vector-like objects of the same length as \code{from}, \code{to}, and \code{membership}.}
}
\value{
A \code{FuzzyHits} object.
}
\description{
The \code{FuzzyHits} class extends the \code{\link[=Hits-class]{Hits}} class to represent hits that are associated with different grades of membership in the interval \code{[0,1]}.
}
\details{
This class does not define any additional slot to the \code{Hits} class.
However, this class defines additional validity checks to ensure that every relation stored in a \code{FuzzyHits} are associated with a numeric membership funtion in the interval \code{[0,1]}.
}
\examples{
# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
membership <- c(0, 0.1, 0.2, 0.3, 0.6, 0.8)

fh <- FuzzyHits(from, to, membership, 7, 15)
fh
}
\seealso{
\code{\link[=Hits-class]{Hits}}, \code{\link[=FuzzySets-class]{FuzzySets}}
}
