% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R,
%   R/FuzzyHits-class.R
\docType{methods}
\name{FuzzyHits-methods}
\alias{FuzzyHits-methods}
\alias{membership}
\alias{membership<-}
\alias{membership,FuzzyHits-method}
\alias{membership<-,FuzzyHits-method}
\title{Methods for \code{FuzzyHits} Objects}
\usage{
membership(object)

membership(object) <- value

\S4method{membership}{FuzzyHits}(object)

\S4method{membership}{FuzzyHits}(object) <- value
}
\arguments{
\item{object}{An object of class inheriting from \link[=FuzzyHits-class]{FuzzyHits}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Accessors'.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\linkS4class{FuzzyHits}} class.
The \code{FuzzyHits} class directly extends \code{\linkS4class{Hits}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{object} represents an object of class inheriting from \code{\linkS4class{FuzzyHits}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{membership(object)} returns a \code{numeric} vector of membership function for each relation.
}

\examples{
# Constructor ----

from <- c(5, 2, 3, 3, 3, 2)
to <- c(11, 15, 5, 4, 5, 11)
membership <- c(0, 0.1, 0.2, 0.3, 0.6, 0.8)

fh <- FuzzyHits(from, to, membership, 7, 15)
fh

# Accessors ----

membership(fh)
fh1 <- fh
membership(fh1)[1] <- 0
}
\seealso{
\code{\linkS4class{FuzzyHits}}, \code{\linkS4class{FuzzySets}}.
}
\author{
Kevin Rue-Albrecht
}
