% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{FuzzySets-class}
\alias{FuzzySets-class}
\alias{FuzzySets}
\title{FuzzySets Class}
\usage{
FuzzySets(relations = DataFrame(element = character(0), set =
  character(0), membership = numeric(0)), ...)
}
\arguments{
\item{relations}{\code{\linkS4class{DataFrame}}.
At least 3 columns that provide mapping relationships between \code{"element"} and \code{"set"}, with \code{"membership"} function in the range \code{[0,1]}.
Additional columns are taken as relation metadata.}

\item{...}{Arguments passed to the \code{\link[=BaseSets]{BaseSets()}} constructor and other functions.}
}
\value{
A \code{FuzzySets} object.
}
\description{
The \code{FuzzySets} class extends the \code{\link{BaseSets}} class to implement a container that also describe different grades of membership in the interval \code{[0,1]}.
}
\details{
This class does not define any additional slot to the \code{BaseSets} class.
However, this class defines additional validity checks to ensure that every relation stored in a \code{FuzzySets} are associated with a numeric membership funtion in the interval \code{[0,1]}.
}
\examples{
# Constructor ----

# Visually intuitive definition of sets, elements, and membership
sets <- list(
  set1=c("A"=0.1, "B"=0.2),
  set2=c("B"=0.3, "C"=0.4, "D"=0.5),
  set3=c("E"=0.8))

# unlist the set names
unlistSets <- rep(names(sets), lengths(sets))
# unlist the element names
unlistElements <- unlist(sapply(sets, names))
# unlist the membership values
unlistMembership <- unlist(sets)

# Reformat as a table
relations <- DataFrame(
  element=unlistElements,
  set=unlistSets,
  membership=unlistMembership
)

fs <- FuzzySets(relations=relations)

# Subsetting ----

fs1 <- subset(fs, set == "set1" | membership > 0.5)

# Coercing ----

# to list (gene sets)
ls1 <- as(fs, "list")
# to matrix (continuous membership)
m1 <- as(fs, "matrix")
# to matrix (multiple observations)
mm1 <- as.matrix(fs, fun.aggregate=min)

# Getters/Setters ----

membership(fs)

fs1 <- fs
membership(fs1) <- runif(length(fs1))
}
\seealso{
\code{\link[=BaseSets-class]{BaseSets}}, \code{\link[=FuzzyHits-class]{FuzzyHits}}, \code{\link{FuzzySets-methods}}.
}
\author{
Kevin Rue-Albrecht
}
