% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/FuzzySet-class.R
\docType{class}
\name{FuzzySets-class}
\alias{FuzzySets-class}
\alias{FuzzySets}
\alias{as.list.FuzzySets}
\title{FuzzySets Class}
\usage{
FuzzySets(..., membership)

\method{as.list}{FuzzySets}(x, ...)
}
\arguments{
\item{...}{Arguments to pass to the \code{\link[=BaseSets]{BaseSets()}} constructor.}

\item{membership}{Numeric. Vector of membership in the range \code{[0,1]}}

\item{x}{An object that inherits from \code{FuzzySets}.}
}
\value{
A \code{FuzzySets} object.
}
\description{
The \code{FuzzySets} class extends the \code{\link{BaseSets}} class to implement a container that also describe different grades of membershipin the interval \code{[0,1]}.
}
\section{Slots}{

\describe{
\item{\code{membership}}{numeric. Membership function.}
}}

\examples{
# Constructor ----

# Visually intuitive definition of sets
sets <- list(
  set1=c("A", "B"),
  set2=c("C", "D"),
  set3=c("E"))

# Reformat as a table
map <- DataFrame(
  element=unlist(sets),
  set=rep(names(sets), lengths(sets))
)

# Generate random values for the membership function
membership <- runif(nrow(map))

fs <- FuzzySets(map=map, membership=membership)

# Subsetting ----

fs1 <- subset(fs, set == "set1" | membership > 0.5)

# Coercing to list ----
ls1 <- as(fs, "list")
}
