% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R, R/GOSets-class.R
\docType{methods}
\name{GOSets-methods}
\alias{GOSets-methods}
\alias{evidence,GOSets-method}
\alias{evidence<-,GOSets-method}
\alias{ontology,GOSets-method}
\alias{ontology<-,GOSets-method}
\title{Methods for \code{GOSets} Objects}
\usage{
\S4method{evidence}{GOSets}(object)

\S4method{evidence}{GOSets}(object) <- value

\S4method{ontology}{GOSets}(object)

\S4method{ontology}{GOSets}(object) <- value
}
\arguments{
\item{object}{An object of class inheriting from \code{\linkS4class{GOSets}}.}

\item{value}{An object of a class specified in the S4 method signature or as outlined in 'Accessors'.}
}
\description{
This page documents the S4 generics and methods defined for objects inheriting of the \code{\linkS4class{GOSets}} class.
The \code{GOSets} class directly extends \code{\linkS4class{Sets}} and thus inherits of all methods defined for the parent class.
In the usage below, \code{object} represents an object of class inheriting from \code{\linkS4class{GOSets}},
and \code{value} is an object of a class specified in the S4 method signature or as outlined in 'Accessors'.
}
\section{Accessors}{

\code{evidence(object)} returns a factor indicating the evidence code for each relation.


\code{ontology(object)} returns a factor indicating the ontology code for each relation.
}

\examples{
# Constructor ----

# Fetch a sample of GO annotations
library(org.Hs.eg.db)
go_sets <- import(org.Hs.egGO)
relations <- as.data.frame(head(go_sets))

gs <- GOSets(relations)

# Accessors ----

evidence(gs)
gs1 <- gs
evidence(gs1)[1] <- "EXP"
ontology(gs)
gs1 <- gs
ontology(gs1)[1] <- "CC"
}
\seealso{
\code{\linkS4class{GOSets}}, \code{\link{Sets-methods}}.
}
\author{
Kevin Rue-Albrecht
}
