% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/IdVector-class.R
\docType{class}
\name{IdVector-class}
\alias{IdVector-class}
\alias{IdVector}
\alias{EntrezIdVector-class}
\alias{EntrezIdVector}
\alias{ids}
\alias{ids<-}
\alias{ids,IdVector-method}
\alias{ids<-,IdVector-method}
\alias{names,IdVector-method}
\alias{names<-,IdVector-method}
\alias{length,IdVector-method}
\alias{[,IdVector,ANY,ANY,ANY-method}
\alias{[,IdVector-method}
\title{IdVector Class}
\usage{
IdVector(ids = character(0))

EntrezIdVector(ids)

ids(x)

ids(x) <- value

\S4method{ids}{IdVector}(x)

\S4method{ids}{IdVector}(x) <- value

\S4method{names}{IdVector}(x)

\S4method{names}{IdVector}(x) <- value

\S4method{length}{IdVector}(x)

\S4method{[}{IdVector,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{ids}{character. Entrez gene identifiers.}

\item{x}{An object that inherits from \code{IdVector}.}

\item{value}{An object of a class specified in the S4 method signature.}

\item{i}{index specifying elements to extract or replace}

\item{j, ..., drop}{Ignored}
}
\value{
A \code{IdVector} object.
}
\description{
The \code{IdVector} class extends the \code{\link{Vector}} class to implement a container that hold a vector of Entrez gene character identifiers.
}
\section{Slots}{

\describe{
\item{\code{ids}}{character. Entrez gene identifiers.}
}}

\examples{
# Constructor ----

tv <- IdVector(ids=head(LETTERS, 6))
mcols(tv) <- DataFrame(row.names = ids(tv), field1=runif(length(tv)))

# Subsetting ----

tv1 <- tv[1:5]

# Constructor ----

library(org.Hs.eg.db)
ev <- EntrezIdVector(keys(org.Hs.eg.db))

}
\seealso{
\code{\link{Vector}}

\code{\link{IdVector}}
}
