% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{IdVector-class}
\alias{IdVector-class}
\alias{IdVector}
\alias{EntrezIdVector-class}
\alias{EntrezIdVector}
\title{IdVector Class}
\usage{
IdVector(ids = character(0))

EntrezIdVector(ids)
}
\arguments{
\item{ids}{character. Identifiers.}
}
\value{
An \code{IdVector} object.
}
\description{
The \code{IdVector} class extends the \code{\link[=Vector-class]{Vector}} class to implement a container that hold a vector of character identifiers.
Subclasses of \code{IdVector} may be defined to enable method dispatch according to the nature of the identifiers (e.g., ENTREZ gene, Gene Ontology term).
}
\section{Slots}{

\describe{
\item{\code{ids}}{character. Identifiers.}
}}

\examples{
# Constructor ----

iv <- IdVector(ids=head(LETTERS, 6))
mcols(iv) <- DataFrame(row.names = ids(iv), field1=runif(length(iv)))
iv

# Subsetting ----

iv[1:5]

# Identifiers/Names ----

ids(iv)
names(iv)

# EntrezIdVector ----

library(org.Hs.eg.db)
eiv <- EntrezIdVector(keys(org.Hs.eg.db))
eiv
# EntrezIdVector ----

library(org.Hs.eg.db)
eiv <- EntrezIdVector(keys(org.Hs.eg.db))
eiv
}
\seealso{
\code{\link[=Vector-class]{Vector}}
}
