% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics.R,
%   R/IdVector-class.R
\docType{class}
\name{IdVector-class}
\alias{IdVector-class}
\alias{IdVector}
\alias{EntrezIdVector-class}
\alias{EntrezIdVector}
\alias{id}
\alias{id<-}
\alias{id,IdVector-method}
\alias{id<-,IdVector-method}
\alias{length,IdVector-method}
\alias{[,IdVector-method}
\title{IdVector Class}
\usage{
IdVector(id = character(0))

EntrezIdVector(id)

id(x)

id(x) <- value

\S4method{id}{IdVector}(x)

\S4method{id}{IdVector}(x) <- value

\S4method{length}{IdVector}(x)

\S4method{[}{IdVector}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{id}{character. Entrez gene identifiers.}

\item{x}{An object that inherits from \code{IdVector}.}

\item{value}{An object of a class specified in the S4 method signature.}

\item{i}{index specifying elements to extract or replace}

\item{j, ..., drop}{Ignored}
}
\value{
A \code{IdVector} object.
}
\description{
The \code{IdVector} class extends the \code{\link{Vector}} class to implement a container that hold a vector of Entrez gene character identifiers.
}
\section{Slots}{

\describe{
\item{\code{id}}{character. Entrez gene identifiers.}
}}

\examples{
# Constructor ----

tv <- IdVector(id=rep(head(LETTERS, 3), each=2))

# Subsetting ----

tv1 <- tv[1:5]

# EntrezIdVector ----

library(org.Hs.eg.db)
ev <- EntrezIdVector(keys(org.Hs.eg.db))

}
\seealso{
\code{\link{Vector}}

\code{\link{IdVector}}
}
