% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\docType{class}
\name{GMTFile-class}
\alias{GMTFile-class}
\alias{GMTFile}
\alias{import.gmt,ANY-method}
\alias{import,GMTFile,ANY,ANY-method}
\alias{import}
\alias{export.gmt}
\alias{export}
\alias{export,BaseSets,GMTFile,ANY-method}
\title{Import and export}
\usage{
GMTFile(resource)

\S4method{import.gmt}{ANY}(con, ...)

\S4method{import}{GMTFile,ANY,ANY}(con, format, text, ...)

export.gmt(object, con, ...)

\S4method{export}{BaseSets,GMTFile,ANY}(object, con, format, ...)
}
\arguments{
\item{resource}{For \code{GMTFile()}, the .gmt file that will be imported in.}

\item{con}{The connection from which data is loaded or to which data is saved.
If this is a character vector, it is assumed to be a filename
and a corresponding file connection is created and then closed after exporting the object.
If a \code{\linkS4class{RTLFile}} derivative, the data is loaded from or saved to the underlying resource.}

\item{...}{Parameters to pass to the format-specific method.}

\item{format, text}{Arguments defined in the \code{\link[rtracklayer:import]{rtracklayer::import()}} generic. Currently ignored.}

\item{object}{An object that inherits from TODO}
}
\value{
For \code{GMTFile()}, an object representing the path to a .gmt file on disk
}
\description{
The functions \code{import} and \code{export} load and save objects from and to particular file formats.
The \code{unisets} package aims to implement support for a number of annotation and sequence formats.
Currently, GMT files are supported.
}
\examples{
# Example file ----

gmt <- system.file(package="unisets", "extdata", "example.gmt")

# Import ----

bs <- import(gmt)

# Export ----

tf <- tempfile(fileext = ".gmt")
export(bs, tf)
}
