% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.r
\docType{class}
\name{GMTFile-class}
\alias{GMTFile-class}
\alias{GMTFile}
\title{GMTFile Class}
\usage{
GMTFile(resource)
}
\arguments{
\item{resource}{For \code{GMTFile()}, the .gmt file that will be imported in.}
}
\value{
For \code{GMTFile()}, an object representing the path to a .gmt file on disk
}
\description{
GMTFile Class
}
