% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniset-package.R
\docType{package}
\name{unisets-package}
\alias{unisets}
\alias{unisets-package}
\title{unisets: Collection of Classes to Store Gene Sets}
\description{
Classes to describe relationships between elements and sets,
    with an emphasis on gene sets.
    Slots are available to store element and set metadata.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinrue/unisets}
  \item \url{http://kevinrue.github.io/unisets}
  \item Report bugs at \url{https://github.com/kevinrue/unisets/issues}
}

}
\author{
\strong{Maintainer}: Kevin Rue-Albrecht \email{kevin.rue-albrecht@kennedy.ox.ac.uk} (0000-0003-3899-3872)

Authors:
\itemize{
  \item Robert Amezquita
}

}
\keyword{internal}
